Schedule nandang

+++ application/controllers/api/C_api.php
@@ -33,7 +33,7 @@ class C_api extends CI_Controller {
         $key = "UAP)(*";
         $data_arr = (array) $this->jwt->decode($token,$key);
 
-        $result = $this->m_api->__getKurikulumByYear($data_arr['year'],$data_arr['ProdiID']);
+        $result = $this->m_api->__getKurikulumByYear($data_arr['SemesterSearch'],$data_arr['year'],$data_arr['ProdiID']);
 
         return print_r(json_encode($result));
     }
@@ -220,8 +220,7 @@ class C_api extends CI_Controller {
                 'CurriculumID' => $insert['CurriculumID'],
                 'ProdiID' => $insert['ProdiID'],
                 'EducationLevelID' => $insert['EducationLevelID'],
-                'MKID' => $insert['MKID'],
-                'MKCode' => $insert['MKCode']);
+                'MKID' => $insert['MKID']);
             $this->db->select('Semester');
             $dataSmt = $this->db->get_where('db_academic.curriculum_details', $where)->result_array();
 
@@ -364,9 +363,9 @@ class C_api extends CI_Controller {
         $data_arr = (array) $this->jwt->decode($token,$key);
 
         if(count($data_arr)>0){
-            $dataForm = (array) $data_arr['dataForm'];
-            if($data_arr['action']=='add'){
 
+            if($data_arr['action']=='add'){
+                $dataForm = (array) $data_arr['dataForm'];
                 // Cek
                 $check = $this->db->get_where('db_academic.semester',array('YearCode'=>$dataForm['YearCode']))
                     ->result_array();
@@ -385,14 +384,26 @@ class C_api extends CI_Controller {
                     return print_r($insert_id);
                 }
 
-            } else if($data_arr['action']=='edit'){
+            }
+            else if($data_arr['action']=='edit'){
+                $dataForm = (array) $data_arr['dataForm'];
                 $this->db->where('ID', $data_arr['ID']);
                 $this->db->update('db_academic.semester',$dataForm);
                 return print_r(1);
-            } else if($data_arr['action']=='delete'){
+            }
+            else if($data_arr['action']=='delete'){
                 $this->db->where('ID', $data_arr['ID']);
                 $this->db->delete('db_academic.semester');
                 return print_r(1);
+            }
+            else if($data_arr['action']=='read'){
+
+                $data = $this->db->order_by('ID', 'DESC')
+                    ->get_where('db_academic.semester',array('IsSemesterAntara'=>'0'))
+                    ->result_array();
+
+                return print_r(json_encode($data));
+
             }
         }
 
@@ -711,7 +722,7 @@ class C_api extends CI_Controller {
             }
             else if($data_arr['action']=='ReadSemesterActive'){
                 $formData = (array) $data_arr['formData'];
-                $data = $this->m_api->getSemesterActive($formData['ProdiID']);
+                $data = $this->m_api->getSemesterActive($formData['CurriculumID'],$formData['ProdiID'],$formData['Semester']);
                 return print_r(json_encode($data));
             }
         }
@@ -725,18 +736,21 @@ class C_api extends CI_Controller {
         if(count($data_arr)>0) {
             if ($data_arr['action'] == 'add') {
                 $formData = (array) $data_arr['formData'];
+                $this->db->insert('db_academic.course_offerings',$formData);
+                $insert_id = $this->db->insert_id();
+                return print_r($insert_id);
+            }
+            else if($data_arr['action']=='edit'){
+                $formData = (array) $data_arr['formData'];
 
-                for($i=0;$i<count($formData);$i++){
-                    $dataInsert = (array) $formData[$i];
-                    $this->db->insert('db_academic.course_offerings',$dataInsert);
-//                    $insert_id = $this->db->insert_id();
-                }
+                $this->db->where('ID', $data_arr['OfferID']);
+                $this->db->update('db_academic.course_offerings',$formData);
 
-                return print_r(1);
+                return print_r($data_arr['OfferID']);
             }
             else if($data_arr['action']=='read'){
                 $formData = (array) $data_arr['formData'];
-                $data = $this->m_api->getAllCourseOfferings($formData['SemesterID'],$formData['ProdiID']);
+                $data = $this->m_api->getAllCourseOfferings($formData['SemesterID'],$formData['CurriculumID'],$formData['ProdiID'],$formData['Semester']);
                 return print_r(json_encode($data));
             }
             else if($data_arr['action']=='readgabungan'){
@@ -766,10 +780,42 @@ class C_api extends CI_Controller {
                 }
             }
             else if($data_arr['action']=='delete'){
-                $this->db->where('ID', $data_arr['ID']);
-                $this->db->delete('db_academic.course_offerings');
 
-                return print_r(1);
+                $query = $this->db->get_where('db_academic.course_offerings', array('ID' => $data_arr['OfferID']), 1)->result_array();
+
+                if(count($query)>0){
+                    $Arr_CDID = json_decode($query[0]['Arr_CDID']);
+
+                    if(count($Arr_CDID)>1){
+                        $result = [];
+                        if (($key = array_search($data_arr['CDID'], $Arr_CDID)) !== false) {
+                            for($a=0;$a<count($Arr_CDID);$a++){
+                                if($a!=$key){
+                                    array_push($result,$Arr_CDID[$a]);
+                                }
+                            }
+                        }
+
+                        $this->db->set('Arr_CDID', json_encode($result));
+                        $this->db->where('ID', $data_arr['OfferID']);
+                        $this->db->update('db_academic.course_offerings');
+
+                        return print_r(1);
+
+
+                    } else if(count($Arr_CDID)==1){
+                        $this->db->where('ID', $data_arr['OfferID']);
+                        $this->db->delete('db_academic.course_offerings');
+                        return print_r(1);
+
+                    }
+
+
+//                    print_r(json_encode($r));
+
+
+                }
+
             }
         }
     }
